разделил сдвоенный класс сервиса
